% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_nces.R
\name{cut_by_n}
\alias{cut_by_n}
\title{Discretize numeric variable}
\usage{
cut_by_n(x, group, n, type)
}
\arguments{
\item{x}{numeric variable}

\item{group}{discrete variable}

\item{n}{number of bins for variable x (by group)}

\item{type}{quantile or cut}
}
\description{
Discretize numeric variable
}
