% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-text-contrast.R
\docType{data}
\name{geom_text_contrast}
\alias{geom_text_contrast}
\alias{GeomTextContrast}
\title{Automatic color selection for text for optimal contrast}
\usage{
geom_text_contrast(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  parse = FALSE,
  nudge_x = 0,
  nudge_y = 0,
  check_overlap = FALSE,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  light = "#FFFFFF",
  dark = "#000000",
  type = "YIQ"
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function. Cannot be jointy specified with
\code{nudge_x} or \code{nudge_y}.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{parse}{If \code{TRUE}, the labels will be parsed into expressions and
displayed as described in \code{?plotmath}.}

\item{nudge_x, nudge_y}{Horizontal and vertical adjustment to nudge labels by.
Useful for offsetting text from points, particularly on discrete scales.
Cannot be jointly specified with \code{position}.}

\item{check_overlap}{If \code{TRUE}, text that overlaps previous text in the
same layer will not be plotted. \code{check_overlap} happens at draw time and in
the order of the data. Therefore data should be arranged by the label
column before calling \code{geom_text()}. Note that this argument is not
supported by \code{geom_label()}.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{light}{hexcode for light text option}

\item{dark}{hexcode for dark text option}

\item{type}{YIQ or Luminance}
}
\description{
These geoms are based on \code{\link[ggplot2]{geom_text}} and
\code{\link[ggplot2]{geom_label}}. See the documentation for those
functions for more details. Differences from those functions are noted
here.
}
\details{
Text with black or white text ...
}
\examples{
library(ggplot2)

ggplot(mpg, aes(x= class, fill = class)) +
  geom_bar() +
  geom_text_contrast(aes(label = class, y = 1), hjust = 0) +
  scale_fill_nces(palette=1) +
  theme_nces() +
  theme(legend.position = "none") +
  coord_flip()

  ggplot(mpg, aes(x= class, fill = class)) +
  geom_bar() +
  geom_text_contrast(aes(label = class, y = 1), type = "Luminance", hjust = 0) +
  scale_fill_nces(palette=1) +
  theme_nces() +
  theme(legend.position = "none") +
  coord_flip()


ggplot(mpg, aes(x= class, fill = class)) +
  geom_bar() +
  geom_text_contrast(aes(label = class, y = 1), hjust = 0) +
  scale_fill_manual(values = c("#d54799", "#394ba0", "#88c6ed",
            "#009f75", "#82c341", "#fff000", "#faa31b")) +
  theme_nces() +
  theme(legend.position = "none") +
  coord_flip()

ggplot(mpg, aes(x= class, fill = class)) +
  geom_bar() +
  geom_text_contrast(aes(label = class, y = 1), type = "Luminance", hjust = 0) +
  scale_fill_manual(values = c("#d54799", "#394ba0", "#88c6ed",
            "#009f75", "#82c341", "#fff000", "#faa31b")) +
  theme_nces() +
  theme(legend.position = "none") +
  coord_flip()

}
\keyword{datasets}
