% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nces_pal.R
\name{nces_palette}
\alias{nces_palette}
\title{Helper function}
\usage{
nces_palette(type = "seq", palette = 1, direction = 1)
}
\arguments{
\item{type}{type of color scheme, one of 'seq' (sequential) or 'div' (divergent)}

\item{palette}{integer specifying the palette, number of 1:6 for sequential colors and 1:3 for divergent}

\item{direction}{order of mapping of variable values to color values, either positive or negative value}
}
\description{
Provide access to one of the pre-defined nces color schemes
}
\examples{
nces_palette()
}
