% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nces_pals}
\alias{nces_pals}
\title{List of NCES color palettes}
\format{
A list of sequential and divergent color schemes
}
\usage{
nces_pals
}
\description{
List of NCES color palettes
}
\keyword{datasets}
