% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nces_pal.R
\name{pals_nces}
\alias{pals_nces}
\title{Palette options}
\usage{
pals_nces()
}
\value{
\code{character} vector containing names of available palettes.
}
\description{
Palette options
}
\examples{
pals_nces()
}
