% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-scales.R
\name{scale_fill_nces}
\alias{scale_fill_nces}
\title{Scales using IES colors}
\usage{
scale_fill_nces(
  ...,
  type = "seq",
  palette = 1,
  direction = -1,
  aesthetics = "fill"
)
}
\description{
Scales using IES colors
}
\examples{
dframe <- data.frame(x = sample(2,100, replace=TRUE))
dframe \%>\%
  ggplot(aes(x = x, fill=factor(x))) +
  geom_bar() +
  scale_fill_nces()
}
