% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-scales.R
\name{scale_colour_nces}
\alias{scale_colour_nces}
\alias{scale_color_nces}
\alias{scale_fill_nces}
\title{Scales using IES colors}
\usage{
scale_colour_nces(
  ...,
  type = "seq",
  palette = 1,
  direction = 1,
  aesthetics = "colour"
)

scale_color_nces(
  ...,
  type = "seq",
  palette = 1,
  direction = 1,
  aesthetics = "colour"
)

scale_fill_nces(
  ...,
  type = "seq",
  palette = 1,
  direction = -1,
  aesthetics = "fill"
)
}
\arguments{
\item{...}{Other arguments passed on to \code{\link[ggplot2:discrete_scale]{discrete_scale()}}, \code{\link[ggplot2:continuous_scale]{continuous_scale()}},
or \code{\link[ggplot2:binned_scale]{binned_scale()}}, for \code{brewer}, \code{distiller}, and \code{fermenter} variants
respectively, to control name, limits, breaks, labels and so forth.}

\item{type}{One of seq (sequential), div (diverging) or qual (qualitative)}

\item{palette}{If a string, will use that named palette. If a number, will index into
the list of palettes of appropriate \code{type}. The list of available palettes can found
in the Palettes section.}

\item{direction}{Sets the order of colours in the scale. If 1, the default,
colours are as output by \code{\link[RColorBrewer:ColorBrewer]{RColorBrewer::brewer.pal()}}. If -1, the
order of colours is reversed.}

\item{aesthetics}{Character string or vector of character strings listing the
name(s) of the aesthetic(s) that this scale works with. This can be useful, for
example, to apply colour settings to the \code{colour} and \code{fill} aesthetics at the
same time, via \code{aesthetics = c("colour", "fill")}.}
}
\description{
Scales using IES colors

Scales using IES colors
}
\examples{
library(tidyverse)
dframe <- data.frame(x = sample(2,100, replace=TRUE))
dframe \%>\%
  ggplot(aes(x = x, fill=factor(x))) +
  geom_bar() +
  scale_fill_nces() +
  theme_bw()
}
