% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-text-contrast.R
\name{text_color}
\alias{text_color}
\title{Light or dark color selection}
\usage{
text_color(
  colors,
  light = "#FFFFFF",
  dark = "#000000",
  type = c("YIQ", "Luminance")
)
}
\arguments{
\item{colors}{vector of background colors}

\item{light}{hexcode for light text option}

\item{dark}{hexcode for dark text option}

\item{type}{YIQ or Luminance}
}
\value{
vector of darks and lights
}
\description{
Light or dark color selection
}
\examples{
colors <- c("#ef4444", "#faa31b", "#fff000", "#82c341",
            "#009f75", "#88c6ed", "#394ba0", "#d54799")

text_color(colors)
}
