% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-nces.R
\name{theme_nces}
\alias{theme_nces}
\title{Theme for NCES}
\usage{
theme_nces(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}
}
\description{
Themes set the general aspect of the plot such as the colour of the
background, gridlines, the size and colour of fonts.
\code{theme_nces} provides access to the regular ggplot2 theme that is been
fine-tuned for NCES.
}
\examples{
library(ggplot2)

ggplot(mpg, aes(class)) +
  geom_bar() +
  theme_nces()

}
